improve data schema, ELN annotations for all quantities
